Get NSF-style download, email, and codebase metrics
These scripts all require AFS access (except extract_git_data
and extract_version_history, which require an up-to-date
CONDOR_SRC git repo, so might as well live adjacent to AFS).

Config should be placed in ~/.weekly-afs-metrics.config.sh
following the instructions in weekly-afs-metrics-update.sh.
Suggested values for the config file will be placed in the wiki.

weekly-afs-metrics-update.sh is the wrapper script around all
of the other scripts in here, and will create a Python 3 virtualenv
containing the requests library (needed for querying the RT API).